backend/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
motor==3.1.1
python-dotenv==0.21.1
pydantic==1.10.4
pymongo==4.3.3
python-multipart==0.0.5
typing-extensions==4.4.0
google-api-python-client==2.88.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
asyncio==3.4.3
aiohttp==3.8.4
celery==5.2.7
redis==4.5.5
